refactor for CLI tool
lays the groundwork but doesn't finish #2
