+++ tools/cli/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "deploy-cli"
+name = "avs-toolkit-cli"
 version = "0.0.0"
 edition = "2021"
 license = "Apache-2.0"
