+++ tools/cli/README.md
