+++ tools/cli/src/config.rs
@@ -1,0 +1,18 @@
+use layer_climb::prelude::*;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Config {
+    pub chains: ChainConfigs,
+    pub faucet: FaucetConfig,
+}
+#[derive(Debug, Deserialize, Serialize)]
+pub struct ChainConfigs {
+    pub local: Option<ChainConfig>,
+    pub testnet: Option<ChainConfig>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct FaucetConfig {
+    pub mnemonic: String,
+}
